+++ packages/typescript-estree/src/parser.ts
@@ -13,19 +13,16 @@ import semver from 'semver';
 import ts from 'typescript';
 import convert from './ast-converter';
 import { convertError } from './convert';
-import { Program } from './estree/spec';
 import { firstDefined } from './node-utils';
-import {
-  ESTreeComment,
-  ESTreeToken,
-  Extra,
-  ParserOptions
-} from './temp-types-based-on-js-source';
+import { TSESTree } from './ts-estree';
+import { Extra, ParserOptions, ParserServices } from './parser-options';
 import { getFirstSemanticOrSyntacticError } from './semantic-errors';
 
-const packageJSON = require('../package.json');
-
-const SUPPORTED_TYPESCRIPT_VERSIONS = packageJSON.devDependencies.typescript;
+/**
+ * This needs to be kept in sync with the top-level README.md in the
+ * typescript-eslint monorepo
+ */
+const SUPPORTED_TYPESCRIPT_VERSIONS = '>=3.2.1 <3.4.0';
 const ACTIVE_TYPESCRIPT_VERSION = ts.version;
 const isRunningSupportedTypeScriptVersion = semver.satisfies(
   ACTIVE_TYPESCRIPT_VERSION,
@@ -49,7 +46,6 @@ function getFileName({ jsx }: { jsx?: boolean }) {
 
 /**
  * Resets the extra config object
- * @returns {void}
  */
 function resetExtra(): void {
   extra = {
@@ -72,9 +68,9 @@ function resetExtra(): void {
 }
 
 /**
- * @param {string} code The code of the file being linted
- * @param {Object} options The config object
- * @returns {{ast: ts.SourceFile, program: ts.Program} | undefined} If found, returns the source file corresponding to the code and the containing program
+ * @param code The code of the file being linted
+ * @param options The config object
+ * @returns If found, returns the source file corresponding to the code and the containing program
  */
 function getASTFromProject(code: string, options: ParserOptions) {
   return firstDefined(
@@ -93,9 +89,9 @@ function getASTFromProject(code: string, options: ParserOptions) {
 }
 
 /**
- * @param {string} code The code of the file being linted
- * @param {Object} options The config object
- * @returns {{ast: ts.SourceFile, program: ts.Program} | undefined} If found, returns the source file corresponding to the code and the containing program
+ * @param code The code of the file being linted
+ * @param options The config object
+ * @returns If found, returns the source file corresponding to the code and the containing program
  */
 function getASTAndDefaultProject(code: string, options: ParserOptions) {
   const fileName = options.filePath || getFileName(options);
@@ -105,13 +101,13 @@ function getASTAndDefaultProject(code: string, options: ParserOptions) {
 }
 
 /**
- * @param {string} code The code of the file being linted
- * @returns {{ast: ts.SourceFile, program: ts.Program}} Returns a new source file and program corresponding to the linted code
+ * @param code The code of the file being linted
+ * @returns Returns a new source file and program corresponding to the linted code
  */
 function createNewProgram(code: string) {
   const FILENAME = getFileName(extra);
 
-  const compilerHost = {
+  const compilerHost: ts.CompilerHost = {
     fileExists() {
       return true;
     },
@@ -162,10 +158,10 @@ function createNewProgram(code: string) {
 }
 
 /**
- * @param {string} code The code of the file being linted
- * @param {Object} options The config object
- * @param {boolean} shouldProvideParserServices True iff the program should be attempted to be calculated from provided tsconfig files
- * @returns {{ast: ts.SourceFile, program: ts.Program}} Returns a source file and program corresponding to the linted code
+ * @param code The code of the file being linted
+ * @param options The config object
+ * @param shouldProvideParserServices True iff the program should be attempted to be calculated from provided tsconfig files
+ * @returns Returns a source file and program corresponding to the linted code
  */
 function getProgramAndAST(
   code: string,
@@ -275,25 +271,21 @@ function warnAboutTSVersion(): void {
 // Parser
 //------------------------------------------------------------------------------
 
-type AST<T extends ParserOptions> = Program &
+type AST<T extends ParserOptions> = TSESTree.Program &
   (T['range'] extends true ? { range: [number, number] } : {}) &
-  (T['tokens'] extends true ? { tokens: ESTreeToken[] } : {}) &
-  (T['comment'] extends true ? { comments: ESTreeComment[] } : {});
+  (T['tokens'] extends true ? { tokens: TSESTree.Token[] } : {}) &
+  (T['comment'] extends true ? { comments: TSESTree.Comment[] } : {});
 
 interface ParseAndGenerateServicesResult<T extends ParserOptions> {
   ast: AST<T>;
-  services: {
-    program: ts.Program | undefined;
-    esTreeNodeToTSNodeMap: WeakMap<object, any> | undefined;
-    tsNodeToESTreeNodeMap: WeakMap<object, any> | undefined;
-  };
+  services: ParserServices;
 }
 
 //------------------------------------------------------------------------------
 // Public
 //------------------------------------------------------------------------------
 
-export const version: string = packageJSON.version;
+export const version: string = require('../package.json').version;
 
 export function parse<T extends ParserOptions = ParserOptions>(
   code: string,
@@ -342,7 +334,7 @@ export function parse<T extends ParserOptions = ParserOptions>(
    * Convert the TypeScript AST to an ESTree-compatible one
    */
   const { estree } = convert(ast, extra, false);
-  return estree;
+  return estree as AST<T>;
 }
 
 export function parseAndGenerateServices<
@@ -406,7 +398,7 @@ export function parseAndGenerateServices<
    * Return the converted AST and additional parser services
    */
   return {
-    ast: estree,
+    ast: estree as AST<T>,
     services: {
       program: shouldProvideParserServices ? program : undefined,
       esTreeNodeToTSNodeMap:
@@ -423,3 +415,5 @@ export function parseAndGenerateServices<
 
 export { AST_NODE_TYPES } from './ast-node-types';
 export { ParserOptions };
+export { ParserServices };
+export { TSESTree };