+++ packages/typescript-estree/tests/lib/comments.ts
@@ -8,10 +8,11 @@
 import { readFileSync } from 'fs';
 import glob from 'glob';
 import { extname } from 'path';
-import { ParserOptions } from '../../src/temp-types-based-on-js-source';
+import { ParserOptions } from '../../src/parser-options';
 import {
   createSnapshotTestBlock,
-  formatSnapshotName
+  formatSnapshotName,
+  isJSXFileType
 } from '../../tools/test-utils';
 
 //------------------------------------------------------------------------------
@@ -35,7 +36,7 @@ describe('Comments', () => {
       range: true,
       tokens: true,
       comment: true,
-      jsx: fileExtension === '.js'
+      jsx: isJSXFileType(fileExtension)
     };
     it(
       formatSnapshotName(filename, FIXTURES_DIR, fileExtension),