+++ packages/typescript-estree/tests/lib/typescript.ts
@@ -8,10 +8,11 @@
 import { readFileSync } from 'fs';
 import glob from 'glob';
 import { extname } from 'path';
-import { ParserOptions } from '../../src/temp-types-based-on-js-source';
+import { ParserOptions } from '../../src/parser-options';
 import {
   createSnapshotTestBlock,
-  formatSnapshotName
+  formatSnapshotName,
+  isJSXFileType
 } from '../../tools/test-utils';
 
 //------------------------------------------------------------------------------
@@ -29,14 +30,16 @@ const testFiles = glob.sync(`${FIXTURES_DIR}/**/*.src.ts`);
 describe('typescript', () => {
   testFiles.forEach(filename => {
     const code = readFileSync(filename, 'utf8');
+    const fileExtension = extname(filename);
     const config: ParserOptions = {
       loc: true,
       range: true,
       tokens: true,
-      errorOnUnknownASTType: true
+      errorOnUnknownASTType: true,
+      jsx: isJSXFileType(fileExtension)
     };
     it(
-      formatSnapshotName(filename, FIXTURES_DIR, extname(filename)),
+      formatSnapshotName(filename, FIXTURES_DIR, fileExtension),
       createSnapshotTestBlock(code, config)
     );
   });